+++ src/Traits/Column/HasDateTimeFormat.php
@@ -8,9 +8,12 @@ trait HasDateTimeFormat
 {
     protected ?string $dateTimeFormat = null;
 
-    public function dateTimeFormat(string $dateTimeFormat): Column
+    protected ?string $timezone = null;
+
+    public function dateTimeFormat(string $dateTimeFormat, string $timezone = null): Column
     {
         $this->dateTimeFormat = $dateTimeFormat;
+        $this->timezone = $timezone;
 
         /** @var \Okipa\LaravelTable\Column $this */
         return $this;
@@ -19,5 +22,10 @@ trait HasDateTimeFormat
     public function getDateTimeFormat(): ?string
     {
         return $this->dateTimeFormat;
+    }
+
+    public function getTimezone(): string
+    {
+        return $this->timezone ?: config('app.timezone');
     }
 }