+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [4.0.6](https://github.com/Okipa/laravel-table/compare/4.0.5...4.0.6)
+
+2022-03-10
+
+* Added ability to set a timezone when formatting a date/time column with the `dateTimeFormat` method:
+    * Updated column `dateTimeFormat` method signature to `dateTimeFormat(string $dateTimeFormat, string $timezone = null): \Okipa\LaravelTable\Column`
+    * If no timezone is set, the default one, defined in `config('app.timezone')` is used
+
 ## [4.0.5](https://github.com/Okipa/laravel-table/compare/4.0.4...4.0.5)
 
 2022-03-01